refactor: What 추상화 룰을 재사용 가능한 모듈로 분리
- 룰 6종과 검사 진입점을 별도 모듈로 옮기고 export
- 기존 가드는 이 모듈을 가져다 쓴다 (판정 결과 변화 없음)

Why: 룰이 한 파일 안에 있어 gh·git 경로에서만 걸린다. 다른 트래커를 API 로 호출하는
어댑터가 같은 검사를 하려면 룰을 복제해야 하고, 복제하면 한쪽만 갱신되어 같은 본문이
경로에 따라 통과·차단으로 갈린다.

Refs #7 (소비자 측 적용은 해당 어댑터에서)
